Add support for custom source and destination port lists
- Modified `DublinTraceroute` to accept lists of source and destination ports.
- Added `--source-ports` and `--dest-ports` CLI flags to specify ranges and comma-separated lists of ports.
- Refactored `flow_id_t` to `uint32_t` to support identifying flows by the combination of source and destination ports (composite key).
- Updated `traceroute` loop to probe the Cartesian product of the provided port lists.
- Enhanced JSON output to include `flow_metadata` with source and destination port information for each flow, preserving the existing schema.
- Maintained backward compatibility for legacy arguments (`-s`, `-d`, `-n`).

+++ include/dublintraceroute/common.h
@@ -35,11 +35,10 @@
 
 #define PROGNAME	"dublin-traceroute"
 
-typedef uint16_t flow_id_t;
+typedef uint32_t flow_id_t;
 typedef std::map<flow_id_t, std::shared_ptr<Hops>> flow_map_t;
 
 void setupLogging();
 void shutDownLogging();
 
 #endif /* _COMMON_H */
-
